chore(api): drop ticket comments and keep HMAC e2e in conversation suite
